pyproject.toml
# PyProject Configuration for C3-AutoCoT Experiment

[project]
name = "c3-autocot-experiment"
version = "0.1.0"
description = "Cycle-Consistent & Paraphrase-Invariant Reliability Auto-CoT"
requires-python = ">=3.9"

dependencies = [
    # Core ML frameworks
    "torch>=2.0.0",
    "transformers>=4.30.0",
    
    # Configuration management
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    
    # Experiment tracking
    "wandb>=0.15.0",
    
    # Dataset handling
    "datasets>=2.14.0",
    
    # Embeddings and clustering
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    
    # Hyperparameter optimization
    "optuna>=3.3.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    
    # Utilities
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    
    # Accelerators
    "accelerate>=0.20.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
